wip: split report renderer helpers
